# [Xamarin.Android.Build.Tasks] remove $(_AndroidResourceNameCaseMap) for performance