--- a/Documentation/release-notes/5188.md
+++ b/Documentation/release-notes/5188.md
@@ -1,0 +1,7 @@
+### Build and deployment performance
+
+  * [GitHub PR 5188](https://github.com/xamarin/xamarin-android/pull/5188):
+    Improve performance by lazily loading an Android resource mapping
+    file that is only used for error messages. This reduced the
+    overall build time by 348ms in a real-world project in an initial
+    clean build or incremental builds with Android resource changes.