--- a/src/Xamarin.Android.Build.Tasks/MSBuild/Xamarin/Android/Xamarin.Android.Aapt2.targets
+++ b/src/Xamarin.Android.Build.Tasks/MSBuild/Xamarin/Android/Xamarin.Android.Aapt2.targets
@@ -102,7 +102,6 @@ Copyright (C) 2011-2012 Xamarin. All rights reserved.
       AndroidConversionFlagFile="$(_AndroidStampDirectory)_ConvertResourcesCases.stamp"
       CustomViewMapFile="$(_CustomViewMapFile)"
       ResourceDirectories="$(MonoAndroidResDirIntermediate);@(_LibraryResourceDirectories)"
-      ResourceNameCaseMap="$(_AndroidResourceNameCaseMap)"
   />
   <Touch Files="$(_AndroidStampDirectory)_ConvertResourcesCases.stamp" AlwaysCreate="True" />
 </Target>
@@ -130,7 +129,6 @@ Copyright (C) 2011-2012 Xamarin. All rights reserved.
       FlatArchivesDirectory="$(_AndroidLibraryFlatArchivesDirectory)"
       ResourcesToCompile="@(_CompileResourcesInputs)"
       ResourceDirectories="$(MonoAndroidResDirIntermediate)"
-      ResourceNameCaseMap="$(_AndroidResourceNameCaseMap)"
       ToolPath="$(Aapt2ToolPath)"
       ToolExe="$(Aapt2ToolExe)"
   />
@@ -161,7 +159,6 @@ Copyright (C) 2011-2012 Xamarin. All rights reserved.
       ContinueOnError="$(DesignTimeBuild)"
       DaemonMaxInstanceCount="$(Aapt2DaemonMaxInstanceCount)"
       DaemonKeepInDomain="$(_Aapt2DaemonKeepInDomain)"
-      ResourceNameCaseMap="$(_AndroidResourceNameCaseMap)"
       AssemblyIdentityMapFile="$(_AndroidLibrayProjectAssemblyMapFile)"
       ImportsDirectory="$(_LibraryProjectImportsDirectoryName)"
       OutputImportDirectory="$(_AndroidLibrayProjectIntermediatePath)"
@@ -223,7 +220,6 @@ Copyright (C) 2011-2012 Xamarin. All rights reserved.
       CompiledResourceFlatFiles="@(_CompiledFlatFiles)"
       DaemonMaxInstanceCount="$(Aapt2DaemonMaxInstanceCount)"
       DaemonKeepInDomain="$(_Aapt2DaemonKeepInDomain)"
-      ResourceNameCaseMap="$(_AndroidResourceNameCaseMap)"
       ResourceDirectories="$(MonoAndroidResDirIntermediate)"
       AssemblyIdentityMapFile="$(_AndroidLibrayProjectAssemblyMapFile)"
       ImportsDirectory="$(_LibraryProjectImportsDirectoryName)"