--- a/src/Xamarin.Android.Build.Tasks/Tasks/Aapt2Compile.cs
+++ b/src/Xamarin.Android.Build.Tasks/Tasks/Aapt2Compile.cs
@@ -41,8 +41,6 @@ namespace Xamarin.Android.Tasks {
 
 		public async override System.Threading.Tasks.Task RunTaskAsync ()
 		{
-			LoadResourceCaseMap ();
-
 			await this.WhenAllWithLock (ResourcesToCompile ?? ResourceDirectories, ProcessDirectory);
 
 			ProcessOutput ();