Reviewing a customer's large solution, `$(_AndroidResourceNameCaseMap)`
contains a ~20,000 character string that is logged and parsed multiple
times throughout the build.

Improvements:

* The `<AndroidComputeResPaths/>` MSBuild task can create a
  `Dictionary<string,string>` directly and store it via
  `RegisterTaskObject`. No need to create a giant string at all.
* `<AndroidComputeResPaths/>` shouldn't log all the output values, as
  it adds hundreds of log messages per `LogDebugTaskItems()` call in
  this customer's project. The surrounding log has enough context, so
  these messages are not needed.
* `MonoAndroidHelper.LoadResourceCaseMap` should only be called when
  the value is actually needed. This is normally for error messages.
* `MonoAndroidHelper.LoadResourceCaseMap` can call
  `GetRegisteredTaskObject` and will no longer need to parse or split
  any strings.
* The `$(_AndroidResourceNameCaseMap)` MSBuild property can be
  completely removed. We no longer need to pass it into various
  MSBuild tasks.

## Results ##

Testing this change in the customer's large solution:

    Before:
       17 ms  ConvertCustomView                          1 calls
      103 ms  AndroidComputeResPaths                    10 calls
     7692 ms  Aapt2Link                                 10 calls
    14288 ms  Aapt2Compile                              10 calls
    After:
       16 ms  ConvertCustomView                          1 calls
       99 ms  AndroidComputeResPaths                    10 calls
     7643 ms  Aapt2Link                                 10 calls
    13994 ms  Aapt2Compile                              10 calls

This should save ~348ms in an initial build of this large solution or
incremental builds with `AndroidResource` changes.

The `.binlog` file produced (which is a compressed format) is also
334,415 bytes smaller.